requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
MarkupSafe==3.0.3
mysql-connector==2.2.9
numpy==2.3.4
openai==2.6.1
openpyxl==3.1.5
pandas==2.3.3
pydantic==2.12.3
pydantic_core==2.41.4
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
